Deduplicate formatDuration, replace padRight with padEnd
